feat: Add guardrail to reset.sh to prevent accidental data loss (#24)

+++ reset.sh
@@ -1,2 +1,7 @@
 #!/usr/bin/env bash
-docker compose down -v
+read -p "Are you sure you want to reset and delete all volumes? [y/N]: " response
+if [[ "$response" =~ ^([yY][eE][sS]|[yY])$ ]]; then
+    docker compose down -v
+else
+    echo "Reset aborted."
+fi
